refactor, now can run with bug
