+++ data/scenes/sphere_dance.json
@@ -18,7 +18,7 @@
 		"exportObj": false,
 		"readPly": true,
 		"plyPath": "E:/codes/try/tiReadPly/test_clip/points",
-		"plyRange":[1,100],
+		"plyRange":[1,2],
 		"readVdb": false,
 		"vdbPath": "D:/CG/meshSequence/sphere_points_v2_vdb/sphere_points_v2_vdb",
 		"vdbObjName":"v",
